renamed records.js store to recordSearch.js
