+++ src/store/index.js
@@ -1,7 +1,7 @@
 import Vue from "vue"
 import Vuex from "vuex"
 import searchFilters from "./searchFilters.js"
-import records from "./records.js"
+import records from "./recordSearch.js"
 import record from "./recordData.js"
 import introspection from "./introspector.js"
 import users from './users.js'
