+++ tests/unit/components/Records/Search/Header/ListController.spec.js
@@ -2,7 +2,7 @@ import {createLocalVue, shallowMount} from "@vue/test-utils";
 import Vuex from "vuex";
 import Vuetify from "vuetify"
 import ListController from "@/components/Records/Search/Header/ListController.vue"
-import recordsStore from "@/store/records.js";
+import recordsStore from "@/store/recordSearch.js";
 
 const localVue = createLocalVue();
 localVue.use(Vuex);
