+++ tests/unit/views/Records/Records.spec.js
@@ -4,7 +4,7 @@ import Vuetify from "vuetify"
 import VueMeta from "vue-meta";
 import Records from "@/views/Records/Records.vue";
 import Client from "@/components/GraphClient/GraphClient.js";
-import records from "@/store/records.js"
+import records from "@/store/recordSearch.js"
 import introspection from "@/store/introspector.js"
 import fakeIntrospection from "@/../tests/fixtures/fakeIntrospection.json"
 import uiController from "@/store/uiController.js"
